[INJIMOB-2265] modify the deserialize method in all classes to handle the key present but value is null scenario and modify invalid input exception to throw error based on field type
description: modifying the deserialize method to parse the data using jsonObject
as the inbuilt kotlin json serialization library CompositeDecoder assigns values based on the index of the field being processed. For any fields not processed (either because the key is missing or the value is null), it will assign null. Therefore, it cannot differentiate between a missing key and a key with a null value, as both result in null.

Signed-off-by: PuBHARGAVI <46226958+PuBHARGAVI@users.noreply.github.com>